requirements.txt
mysql-connector
flask
flask_session
Werkzeug
usaddress
numpy
pymc
geopy
requests
scikit-learn